[Silver II] Title: 나무 자르기, Time: 1288 ms, Memory: 193680 KB -BaekjoonHub

+++ 백준/Silver/2805. 나무 자르기/나무 자르기.java
@@ -22,7 +22,7 @@ public class Main {
 		}
 
 		int h = q.peek(), cnt = 0;
-		long sum = 0L;
+		int sum = 0;
 		while (!q.isEmpty() && sum < m && h > 0) {
 			h--;
 			if (q.peek() <= h) {
